LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/is-expected.txt

Red
Red
Green
Green
Blue
Blue
Red
Blue
Red

PASS Preconditions. 
PASS Invalidate :is() for simple selector arguments. 
PASS Invalidate :is() for compound selector arguments. 
PASS Invalidate :is() for complex selector arguments. 
PASS Invalidate nested :is(). 
FAIL Test specificity of :is(). assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
